First crack at dispensary list
The code first gets all businesses in the state of CO.  I have an API key so we can pull a lot, but I didn't want to get too greedy.  I'm starting with 500,000.  I then narrow down the 500,000 to only businesses opened after 12/31/2013, because recreational mj stores were allowed to open on 1/1/2014.  Then, it loads a csv of all the dispensaries in CO and I merge this with the first dataset on street address.